Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid
Click
ecdsa
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pycryptodome
pylint
python-jose[cryptography]
six
# typed-ast==1.4.2
Werkzeug
wrapt
Flask-Cors
SQLAlchemy
python-dotenv